pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "framework-argb-wallpaper-sync"
version = "0.1.0"
description = "Sync Framework laptop RGB keyboard/fans to current wallpaper"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0"]

[project.scripts]
colour-sync = "framework_argb_wallpaper_sync.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
